* ecoff.c (bfd_debug_section): Fix initialization. 	* elf.c (_bfd_elf_slurp_version_tables): Change maxidx to unsigned, it 	is always a positive integer. Cast away sign mismatch. 	* elf32-mips.c: Fix misleading comment and typo. 	(_bfd_mips_elf_section_from_bfd_section): Remove unused attribute, use 	correct data type. 	* elflink.c: Fix typo. 	(_bfd_elf_create_dynamic_sections): Remove superfluous initialization. 	* ecoffswap.h (ecoff_swap_fdr_in): Cast away sign mismatch.

+++ bfd/ecoffswap.h
@@ -196,7 +196,7 @@ ecoff_swap_fdr_in (abfd, ext_copy, intern)
   intern->adr           = ecoff_get_off (abfd, (bfd_byte *)ext->f_adr);
   intern->rss           = bfd_h_get_32 (abfd, (bfd_byte *)ext->f_rss);
 #if defined (ECOFF_64) || defined (ECOFF_SIGNED_64)
-  if (intern->rss == 0xffffffff)
+  if (intern->rss == (signed long) 0xffffffff)
     intern->rss = -1;
 #endif
   intern->issBase       = bfd_h_get_32 (abfd, (bfd_byte *)ext->f_issBase);
